requirements/local.txt
-r test.txt

# Code analysis tools
coverage==5.0.2
django-coverage-plugin==1.6.0
flake8==3.7.9 # pyup: != 2.6.0
pylint==2.4.4
mccabe==0.6.1

# Debug tools - Note that upgrading django-extensions to 2.1.7 gave error "CommandError:
# Unknown database engine django.contrib.gis.db.backends.postgis" on Ubuntu only.
django-extensions==2.2.5
django-debug-toolbar==2.1
ipdb==0.12.3
Werkzeug==0.16.0

# Build and code maintenance tools
docker==4.1.0
docker-compose==1.25.1
pur==5.3.0

# documentation, TODO remove
Sphinx==2.3.1

# Parse yaml files
pyyaml==5.3